% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clboxplot.R
\name{clboxplot}
\alias{clboxplot}
\title{clboxplot}
\usage{
clboxplot(dataname, factor, factorname)
}
\arguments{
\item{dataname}{: summary dataframe of cowlog observations (generated via cldata)}

\item{factor}{: must be listed as either TRUE or FALSE, if false boxplots of summary distributions are generated, if true boxplots are generated and compared by category (t-tests and anovas)}

\item{factorname}{: (used if factor = TRUE) name of relevant categorical column in summary dataframe}
}
\description{
To visualize and compare the distribution of values, clboxplot generates bosplots of values and additionally can compares distributions across categories via t-tests and anovas.
}
\examples{
clboxplot(dataname = dataframe_round, factor = TRUE, factorname = "round")
}
\keyword{boxplot}
\keyword{cowlog}
