% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cldata.R
\name{cldata}
\alias{cldata}
\title{cldata}
\usage{
cldata(
  pathtofile,
  outputdataname,
  outputzonename,
  factor,
  factorindex,
  factorname
)
}
\arguments{
\item{pathtofile}{: path to folder hosting individual csv's generated by cowlog}

\item{outputdataname}{: string of desired name for output summary dataframe of cowlog observations}

\item{outputzonename}{: string of desired name for output list of zones identified in the cowlog observations}

\item{factor}{: must be listed as either TRUE or FALSE, if true, one categorical factor is incorporated into the resulting summary dataframe}

\item{factorindex}{: (used if factor = TRUE) index in cowlog individual csv's filename (separted by "_") that the factor of interest is located}

\item{factorname}{: (used if factor = TRUE) string desired name for output column in summary dataframe referring to the factor of interest}
}
\description{
cldata generates a summary dataframe based on individual cowlog csv observations. Additionaly it generates a list of zones dientified in the observations.
}
\examples{
cldata(pathtofile = "C:/Users/example_files", outputdataname = "dataframe_round", outputzonename = "list_of_zones", factor = TRUE, factorindex = 3, factorname = "round")
}
\keyword{cowlog}
\keyword{dataframe}
