% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clflag.R
\name{clflag}
\alias{clflag}
\title{clflag}
\usage{
clflag(pathtofile)
}
\arguments{
\item{pathtofile}{: path to folder hosting individual csv's generated by cowlog}
}
\description{
clflag is designed to identify any cowlog csv observation sheets that do may host scorer input errors. It flags csv files that do not include an END entry, that have fewer than entries, and that report negative time differences between entries.
}
\examples{
clflag(pathtofile = "C:/Users/example_files")
}
\keyword{cowlog}
\keyword{flag}
