% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clpie.R
\name{clpie}
\alias{clpie}
\title{clpie}
\usage{
clpie(dataname, zonename, factor, factorname)
}
\arguments{
\item{dataname}{: summary dataframe of cowlog observations (generated via cldata)}

\item{zonename}{: list of zones (generated via cldata)}

\item{factor}{: must be listed as either TRUE or FALSE, if true pie graphs are seperately generated for each category}

\item{factorname}{: (used if factor = TRUE) name of relevant categorical column in summary dataframe}
}
\description{
To visualize where individuals are spending time in the observations, clpie generates a pie graph of proportion time in each zone, and can generate pie graphs per category.
}
\examples{
clpie(dataname = dataframe_round, zonename = list_of_zones, factor = TRUE, factorname = "round")
}
\keyword{cowlog}
\keyword{pie}
