% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clreg.R
\name{clreg}
\alias{clreg}
\title{clreg}
\usage{
clreg(dataname, zonename, factor, factorname)
}
\arguments{
\item{dataname}{: summary dataframe of cowlog observations (generated via cldata)}

\item{zonename}{: list of zones (generated via cldata)}

\item{factor}{:  must be listed as either TRUE or FALSE, if false scatterplots and regresison models of all data are generated, if true scatterplots are colored by category and p values of interaction effects are reported}

\item{factorname}{:(used if factor = TRUE) name of relevant categorical column in summary dataframe}
}
\description{
To determine which metrics correlate with each other and/or show interaction effects of factors, clreg runs linear regression models and generates summary scatterplots.
}
\examples{
clreg(data = dataframe_round, zonename = list_of_zones, factor = TRUE, factorname = "round")
}
\keyword{cowlog}
\keyword{linear}
\keyword{regression}
