% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clseries.R
\name{clseries}
\alias{clseries}
\title{clseries}
\usage{
clseries(pathtofile, zonename, seglength, factor, factorindex, factorname)
}
\arguments{
\item{pathtofile}{: path to folder hosting individual csv's generated by cowlog}

\item{zonename}{: list of zones (generated via cldata)}

\item{seglength}{: desired length (in seconds) of each segment}

\item{factor}{:  must be listed as either true or false, if false one location series graph is averaged acros all data, if true location series graphs are generated per factor category}

\item{factorindex}{: (used if factor = TRUE) index in cowlog individual csv's filename that the factor of interest is located (separted by "_")}

\item{factorname}{: (used if factor = TRUE) name of desired factor to apprear on the location series graphs}
}
\description{
To visualize average movement in the arena over time, clseries splits each observation into ten even segments of the same length and visualizes relative time in zones per segment.
}
\examples{
clseries(pathtofile = "C:/Users/example_files", zonename = zones, seglength = 20, factor = TRUE, factorindex = 3, factorname = "round")
}
\keyword{cowlog}
\keyword{location}
\keyword{series}
\keyword{time}
